requirements.txt
anyio==3.7.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
autopep8==2.0.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
fastapi==0.97.0
fonttools==4.39.4
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
pdfkit==1.0.0
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pydyf==0.6.0
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.7.0
pypdf==3.9.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
starlette==0.27.0
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.6.3
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.1
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
whitenoise==6.5.0
zopfli==0.2.2
boto3==1.28.20
python-dotenv==1.0.0
django-rest-swagger==2.2.0
gunicorn==21.2.0
django-debug-toolbar==4.2.0
django-crontab==0.7.1
retry==0.9.2
